+++ tensorflow_datasets/text/__init__.py
@@ -18,6 +18,7 @@
 from tensorflow_datasets.text.c4 import C4
 from tensorflow_datasets.text.definite_pronoun_resolution import DefinitePronounResolution
 from tensorflow_datasets.text.gap import Gap
+from tensorflow_datasets.text.google_conceptual_captions import GoogleConceptualCaptions
 from tensorflow_datasets.text.glue import Glue
 from tensorflow_datasets.text.imdb import IMDBReviews
 from tensorflow_datasets.text.imdb import IMDBReviewsConfig